+++ src/main/java/org/realityforge/bazel/depgen/config/AliasStrategy.java
@@ -1,9 +1,0 @@
-package org.realityforge.bazel.depgen.config;
-
-public enum AliasStrategy
-{
-  // Alias is generated based on the groupId and artifactId of the artifact
-  GroupIdAndArtifactId,
-  // Alias is generated based on the artifactId of the artifact
-  ArtifactId
-}