+++ src/test/java/org/realityforge/bazel/depgen/model/OptionsModelTest.java
@@ -3,7 +3,7 @@ package org.realityforge.bazel.depgen.model;
 import gir.io.FileUtil;
 import java.nio.file.Path;
 import org.realityforge.bazel.depgen.AbstractTest;
-import org.realityforge.bazel.depgen.config.AliasStrategy;
+import org.realityforge.bazel.depgen.config.NameStrategy;
 import org.realityforge.bazel.depgen.config.GlobalJavaConfig;
 import org.realityforge.bazel.depgen.config.Nature;
 import org.realityforge.bazel.depgen.config.OptionsConfig;
@@ -27,7 +27,7 @@ public class OptionsModelTest
     assertEquals( model.getWorkspaceMacroName(), OptionsConfig.DEFAULT_WORKSPACE_MACRO_NAME );
     assertEquals( model.getTargetMacroName(), OptionsConfig.DEFAULT_TARGET_MACRO_NAME );
     assertEquals( model.getNamePrefix(), OptionsConfig.DEFAULT_NAME_PREFIX );
-    assertEquals( model.getAliasStrategy(), OptionsConfig.DEFAULT_ALIAS_STRATEGY );
+    assertEquals( model.getNameStrategy(), OptionsConfig.DEFAULT_NAME_STRATEGY );
     assertEquals( model.getDefaultNature(), OptionsConfig.DEFAULT_NATURE );
     assertTrue( model.failOnMissingPom() );
     assertTrue( model.failOnInvalidPom() );
@@ -64,7 +64,7 @@ public class OptionsModelTest
     source.setWorkspaceMacroName( "gen_myprj_dependency_rules" );
     source.setTargetMacroName( "gen_myprj_targets" );
     source.setNamePrefix( "myprj_" );
-    source.setAliasStrategy( AliasStrategy.ArtifactId );
+    source.setNameStrategy( NameStrategy.ArtifactId );
     source.setDefaultNature( Nature.J2cl );
     source.setFailOnMissingPom( false );
     source.setFailOnInvalidPom( false );
@@ -84,7 +84,7 @@ public class OptionsModelTest
     assertEquals( model.getWorkspaceMacroName(), "gen_myprj_dependency_rules" );
     assertEquals( model.getTargetMacroName(), "gen_myprj_targets" );
     assertEquals( model.getNamePrefix(), "myprj_" );
-    assertEquals( model.getAliasStrategy(), AliasStrategy.ArtifactId );
+    assertEquals( model.getNameStrategy(), NameStrategy.ArtifactId );
     assertEquals( model.getDefaultNature(), Nature.J2cl );
     assertFalse( model.failOnMissingPom() );
     assertFalse( model.failOnInvalidPom() );